Create create-users.ps1
PowerShell script to automate bulk creation of local Windows user accounts from a CSV file. Useful for IT help desk, sysadmins, and onboarding users in enterprise environments. Edit users.csv with username and password columns, run the script in PowerShell as admin, and all accounts will be created and added to the Users group.